# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "个人博客后端API",
        "title": "Blog API",
        "termsOfService": "https://www.google.com",
        "contact": {
            "name": "www.google.com",
            "url": "https://www.google.com",
            "email": "kakazhang10@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/user": {
            "post": {
                "description": "添加用户",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Blog文档"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "passwd",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "github",
                        "name": "github",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message\": fmt.Sprint(userJson.Username, \" insert sucessfully!\")",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "message\": fmt.Sprint(userJson.Username, \" insert failed!\"),",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/:id": {
            "get": {
                "description": "通过ID查询用户信息",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Blog文档"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "int",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "error\": \"Record not found!\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "data\": user}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}